moved api dto objects to addon packages
